Actually skip when we do skip_all

+++ Rakefile
@@ -37,6 +37,7 @@ def create_links linkables
         when 's' then next
         end
       end
+      next if skip_all
       FileUtils.rm_rf(target) if overwrite || overwrite_all
       `mv "$HOME/.#{file}" "$HOME/.#{file}.backup"` if backup || backup_all
     end
